Added tests for uploading surahs in mushaf data, moved data upload functionality to new service
