+++ quranapp_backend/src/controllers/data_upload.py
@@ -3,7 +3,7 @@ from pydantic import ValidationError
 from sqlalchemy.orm import Session
 
 from src.controllers.dependencies import db_session_dependency, api_key_dependency
-from src.services import mushafs as mushafs_service
+from src.services import data_upload as data_upload_service
 
 data_upload_router = APIRouter(
     prefix="/data-upload",
@@ -20,12 +20,12 @@ def upload_ayah_parts_data(
     if not data_file.filename.endswith(".json"):
         raise HTTPException(detail="Data file must have a .json extension", status_code=status.HTTP_400_BAD_REQUEST)
 
-    service = mushafs_service.AyahPartsDataUploader(db=db)
+    service = data_upload_service.AyahPartsDataUploader(db=db)
     try:
         service.save_data_from_ayah_parts_file(data_file.file)
     except ValidationError as e:
         raise HTTPException(detail=e.errors(), status_code=status.HTTP_400_BAD_REQUEST)
-    except mushafs_service.DataUploadException as e:
+    except data_upload_service.DataUploadException as e:
         raise HTTPException(detail=str(e), status_code=status.HTTP_400_BAD_REQUEST)
     finally:
         data_file.file.close()
@@ -41,12 +41,12 @@ def upload_page_images_data(
             detail="Data file must have a .json extension", status_code=status.HTTP_400_BAD_REQUEST
         )
 
-    service = mushafs_service.PageImagesDataUploader(db=db)
+    service = data_upload_service.PageImagesDataUploader(db=db)
     try:
         service.save_data_from_page_images_file(data_file.file)
     except ValidationError as e:
         raise HTTPException(detail=e.errors(), status_code=status.HTTP_400_BAD_REQUEST)
-    except mushafs_service.DataUploadException as e:
+    except data_upload_service.DataUploadException as e:
         raise HTTPException(detail=str(e), status_code=status.HTTP_400_BAD_REQUEST)
     finally:
         data_file.file.close()
@@ -62,12 +62,12 @@ def upload_surahs_in_mushaf_data(
             detail="Data file must have a .json extension", status_code=status.HTTP_400_BAD_REQUEST
         )
 
-    service = mushafs_service.SurahsInMushafDataUploader(db=db)
+    service = data_upload_service.SurahsInMushafDataUploader(db=db)
     try:
         service.save_data_from_surahs_in_mushaf_file(data_file.file)
     except ValidationError as e:
         raise HTTPException(detail=e.errors(), status_code=status.HTTP_400_BAD_REQUEST)
-    except mushafs_service.DataUploadException as e:
+    except data_upload_service.DataUploadException as e:
         raise HTTPException(detail=str(e), status_code=status.HTTP_400_BAD_REQUEST)
     finally:
         data_file.file.close()
