+++ quranapp_backend/src/services/__init__.py
@@ -2,5 +2,6 @@ import src.services.auth as tokens_service
 import src.services.users as users_service
 import src.services.surahs as surahs_service
 import src.services.mushafs as mushafs_service
+import src.services.data_upload as data_upload_service
 import src.services.mushaf_pages as mushaf_pages_service
 import src.services.utils
