+++ quranapp_backend/tests/test_data/surahs_in_mushaf_upload/surahs_updated.json
@@ -1,0 +1,14 @@
+{
+  "riwayah": "Qaloon",
+  "publisher": "Madina",
+  "surahs": [
+    {
+      "surah_number": 500,
+      "page_number": 6005
+    },
+    {
+      "surah_number": 501,
+      "page_number": 6001
+    }
+  ]
+}
